Instrumentation, test and storage
Changed logging in instrumentation.
Changed `measure_point` name from `__name__` to a string defining the name.
Changed fixture to define a `measure_session`.
Changed fixture to make a tuple representation of it for better testing the db result.

Added instrumentation of `sqlalchemy.orm.Session.commit`.
Added tests for `measure_session`
Added a measurement session identifier.

Split the subcalls of a measurement in sqlite storage in a own table.